refactor(work-creator): move updateStatus orchestration into lib package
Extract the updateStatus function and its helpers (handleWorkflowControlFile,
addWorkflowSuspendLabel, readStatusFile) out of cmd/update_status.go into
lib/status_updater.go as the exported lib.UpdateStatus. cmd/update_status.go
becomes a thin cobra wrapper. No behaviour change; this prepares for the
upcoming 'run' subcommand that will call both lib.WorkCreator.Execute and
lib.UpdateStatus.
